Fixing error on last push
Since we are going to actually create ScriptableObjects, they have to have [CreateAssetMenu(menuName = "Inventory Rune Data")] which allows us to create prefabs of that data in /Assets/ScriptableObjects folder. Error was that menuName was everywhere "Inventory Rune Data" and had to be replaced with each of its respective names.