Updated Inventory and More
Started actually managing data, creating instances of rune items. For now the instance items are being held in a list and there is a dictionary of those items that is being kept in the Canvas>Inventory>Runes. Run a testing script if adding to that list and then to the stack works - it does. You can open the inventory with a key stroke of "Tab". Updated some basic packages, installed Cinemachine camera and currently set it in a basic way to follow the player. Next I will visually represent the runes in the inventory, it's quantity and then develop a pick-up system for runes. If that works, then move to other inventory systems - spells, cards, adjusting cards etc.